fix(react): resolve React Doctor findings for desktop and navigation UI
