+++ .github/workflows/react-doctor.yml
@@ -13,6 +13,8 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
+        with:
+          fetch-depth: 0
       - uses: millionco/react-doctor@main
         with:
           github-token: ${{ secrets.GITHUB_TOKEN }}
